Move routes
Extract TransactionViewer;
Move /transactions to /base64;
